Copter: allow for flight mode UI when FRAME_CLASS=0
this makes the user experience of a first time install nicer for
copters/helis. At the moment the user finds the GCS is not functional,
for example flight mode drop-downs don't display. This can be
confusing as the user doesn't know about the link between displaying
flight modes and the FRAME_CLASS